Move SavedData to a new component folder
